To display ordered items in a page after response on cart
